% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{genera.dati.multicentro}
\alias{genera.dati.multicentro}
\title{Funzione per generare dei dati di uno studio multicentrico}
\usage{
genera.dati.multicentro(numero.centri, numeroCovariate = 1,
  numeroCampioni.x.covariata = 10, noise.intra.X = list(),
  noise.inter.X = list(), intra.power.distorsion = c(),
  inter.power.distorsion = 1)
}
\arguments{
\item{numero.centri}{(numeric) il numero di centri dello studio multicentrico da simulare}

\item{<tutti}{gli altri parametri> fate riferimento all'help della funzione 'generate.set()'}
}
\description{
E' una funzione che genera le covariate e l'outcome, ipotizzando il contributo di diversi centri (tutti con la stessa statistaica)
}
\examples{
\dontrun{

# Multicentrico di due centri, con tre covariate lineari leggermente disturbate con una gaussiana
a<-genera.dati.multicentro(numero.centri = 3, noise.intra.X = list("gaussian"=list("mean"=0,"sd"=0.2)))
 
} 
}

